simple_crypto, simple_proto, fsl_sec: Fix KW Sev8&9 issues
This patch fixes all-around the place the issues identified by KW
with Severity 8&9:
- Comparison of unsigned value against 0 is always true
- Value is Never Used after Assignment
- Loss of Precision
- Loss of Precision during function call

Reviewed-on: http://git.am.freescale.net:8181/19425
Tested-by: Vakul Garg <vakul@freescale.com>

+++ apps/simple_crypto/app/simple_crypto.c
@@ -419,8 +419,8 @@ void set_buf_size(struct test_param *crypto_info)
  * brief	Setup preheader for shared descriptor
  * return	0 on success, otherwise -ve value
  */
-void *setup_preheader(uint32_t shared_desc_len, uint32_t pool_id,
-		      uint32_t pool_buf_size, uint8_t absolute, uint8_t add_buf)
+void *setup_preheader(uint32_t shared_desc_len, uint8_t pool_id,
+		      uint8_t pool_buf_size, uint8_t absolute, uint8_t add_buf)
 {
 	struct preheader_s *prehdr = NULL;
